@@ -1,5 +1,5 @@
 module github.com/alesforz/cryptopals
 
-go 1.21.0
+go 1.22.2
 
 require golang.org/x/sync v0.3.0
